앱 테마 변경, 아이콘 추가
