+++ app/src/main/res/layout/activity_login.xml
@@ -40,7 +40,7 @@
         android:singleLine="true"
         android:textSize="25sp"
         android:textColor="@color/black"
-
+        android:backgroundTint="@color/theme"
         android:textAppearance="@style/TextAppearance.AppCompat.Body1"
         android:hint="아이디" />
 
@@ -67,7 +67,7 @@
         android:inputType="textPassword"
         android:textSize="25sp"
         android:textColor="@color/black"
-
+        android:backgroundTint="@color/theme"
         app:passwordToggleEnabled="true"
         android:textAppearance="@style/TextAppearance.AppCompat.Body1"
         android:hint="비밀번호" />
