+++ app/src/main/res/layout/activity_main.xml
@@ -18,30 +18,12 @@
             android:layout_height="44dp"
             android:background="#FFFFFF"
             android:paddingTop="5dp"
-            android:paddingStart="10dp"
+            android:paddingStart="150dp"
             android:paddingEnd="5dp"
+            app:navigationIcon="@drawable/img_toolbar2"
             app:layout_scrollFlags="scroll|enterAlways"
             app:titleTextColor="@android:color/black"
             app:titleTextAppearance="@style/TextAppearance.AppCompat.Medium"/>
-        <LinearLayout
-            android:layout_width="match_parent"
-            android:orientation="horizontal"
-            android:background="@android:color/white"
-            android:layout_height="wrap_content">
-            <Button
-                android:id="@+id/main_btn_logout"
-                android:layout_width="wrap_content"
-                android:layout_weight="1"
-                android:text="로그아웃"
-                android:layout_height="wrap_content"/>
-            <Button
-                android:id="@+id/main_btn_signout"
-                android:layout_width="wrap_content"
-                android:text="회원탈퇴"
-                android:layout_weight="1"
-                android:layout_height="wrap_content"/>
-        </LinearLayout>
-
 
         <com.google.android.material.tabs.TabLayout
             android:id="@+id/main_tabs"
@@ -52,7 +34,7 @@
             app:tabMode="fixed"
             app:tabIconTint="@android:color/darker_gray"
             app:tabGravity="fill"
-            app:tabIndicatorColor="@android:color/black" />
+            app:tabIndicatorColor="@color/theme" />
 
     </com.google.android.material.appbar.AppBarLayout>
